CAS/requirements.txt
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
pymoo==0.6.1.3
scikit-learn==1.6.1
ngsolve==6.2.2404.post51
netgen-mesher==6.2.2404.post16